Addressed feedback received on 05/16
- Moved `ConsumerPair` and `PartitionMovements` definitions inside `StickyAssignor`
- Moved the schema definition for topic partition list and its serializer/deserializer inside `StickyAssignor`
- Limited the exposure of newly added class members
