+++ clients/src/main/java/org/apache/kafka/common/TopicPartition.java
@@ -75,5 +75,4 @@ public final class TopicPartition implements Serializable {
     public String toString() {
         return topic + "-" + partition;
     }
-
 }
